Refactor data analysis, fix imports
