+++ air-quality-app/client/src/App.css
@@ -33,6 +33,8 @@
   width: 60%;
   border-radius: 1rem;
   opacity: 0.8;
+  margin: 2rem;
+  padding: 2rem;
 }
 
 .Loader {
