+++ air-quality-app/client/src/components/DataAnalysis.tsx
@@ -1,0 +1,25 @@
+import React from "react";
+
+
+export function DataAnalysis(){
+    return (
+        <div className='Lore'>
+          Ziemia i środowisko zmagają się z poważnym problemem zwiększającego się zanieczyszczenia powietrza, wody oraz gleby.  Zanieczyszczenia te są spowodowane w większości działalnością człowieka i mogą mieć fatalne skutki dla flory, fauny i ludzi. Jednymi z najgroźniejszych cząsteczek są tlenki węgla, dwutlenki azotu oraz różne inne cząsteczki zawieszone w powietrzu o średnicy nie większej niż 10 mikrometrów (określane nazwą PM10), pochodzące najczęściej z spalania paliw stałych i ciekłych. <br></br><br></br>
+
+          Jednak poziom zanieczyszczenia powietrza nie jest związany wyłącznie z działalnością człowieka. Ma na niego wpływ wielu czynników takich jak warunki pogodowe. W naszych obliczeniach skupiliśmy się na zależności pomiędzy temperaturą powietrza, prędkością wiatru oraz zachmurzeniem, a jakością powietrza. Nasze pomiary pochodzą z placówki Instytutu Meteorologii i Gospodarki Wodnej w Gdańsku oraz siedmiu placówek dokonujących pomiaru zanieczyszczeń powietrza na terenie Trójmiasta. <br></br><br></br>
+
+          --------wykresy i zdjęcia---------------
+          
+          <br></br><br></br>
+          Jak można zauważyć, największa korelacja występuje pomiędzy temperaturą a zawartością tlenku węgla oraz dwutlenku azotu. Istnieje również zależność między temperaturą a PM10, jednak nie jest tak znacząca. Im temperatura jest niższa, tym powietrze staje się bardziej zanieczyszczone.
+          Połączenie to jest spowodowane spalaniem większej ilości paliw w celu ogrzania mieszkań, biur czy fabryk.
+          <br></br><br></br>
+          
+          Równie znaczącą relacją jest to, że wszystkie zanieczyszczenia występują odwrotnie proporcjonalnie do prędkości wiatru, który im z większą prędkością wieje, tym bardziej rozwiewa zawieszone w nim cząsteczki.
+          <br></br>
+          Nasze obliczenia wskazały brak jednoznacznego powiązania między obecnością szkodliwych substancji w powietrzu a zachmurzeniem.
+          <br></br><br></br>
+          Podsumowując najlepszą porą do spędzania czasu na dworze są ciepłe i wietrzne dni. :)
+        </div>
+    )
+}
