+++ air-quality-app/client/src/components/graphs/PeriodPicker.tsx
@@ -3,8 +3,8 @@
 import React, {ChangeEvent, useState} from "react";
 import { Calendar } from "react-calendar";
 import 'react-calendar/dist/Calendar.css';
-import '../styles/Calendar.css'
-import '../styles/Picker.css'
+import '../../styles/Calendar.css'
+import '../../styles/Picker.css'
 import Select from "react-select";
 import { TimePeriod } from "../../types";
